+++ junos/resource_security_ipsec_vpn.go
@@ -6,6 +6,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/hashicorp/go-cty/cty"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
@@ -48,9 +49,8 @@ func resourceIpsecVpn() *schema.Resource {
 				Computed: true,
 			},
 			"bind_interface_auto": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				ForceNew: true,
+				Type:          schema.TypeBool,
+				Optional:      true,
 				ConflictsWith: []string{"traffic_selector"},
 			},
 			"df_bit": {
@@ -238,32 +238,51 @@ func resourceIpsecVpnRead(ctx context.Context, d *schema.ResourceData, m interfa
 	return nil
 }
 func resourceIpsecVpnUpdate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+	var diagsReturn diag.Diagnostics
 	d.Partial(true)
 	sess := m.(*Session)
 	jnprSess, err := sess.startNewSession()
 	if err != nil {
-		return diag.FromErr(err)
+		return append(diagsReturn, diag.FromErr(err)...)
 	}
 	defer sess.closeSession(jnprSess)
 	sess.configLock(jnprSess)
 	if err := delIpsecVpnConf(d, m, jnprSess); err != nil {
 		sess.configClear(jnprSess)
 
-		return diag.FromErr(err)
+		return append(diagsReturn, diag.FromErr(err)...)
+	}
+	if d.HasChanges("bind_interface_auto") {
+		diagsReturn = append(diagsReturn, diag.Diagnostic{
+			Severity:      diag.Warning,
+			Summary:       "Modifying bind_interface_auto on resource already created has no effect",
+			AttributePath: cty.Path{cty.GetAttrStep{Name: "bind_interface_auto"}},
+		})
+	}
+	if d.HasChanges("bind_interface") {
+		oldInt, _ := d.GetChange("bind_interface")
+		empty := checkInterfaceNC(oldInt.(string), m, jnprSess)
+		if empty == nil {
+			if err := sess.configSet([]string{"delete interfaces " + oldInt.(string)}, jnprSess); err != nil {
+				sess.configClear(jnprSess)
+
+				return append(diagsReturn, diag.FromErr(err)...)
+			}
+		}
 	}
 	if err := setIpsecVpn(d, m, jnprSess); err != nil {
 		sess.configClear(jnprSess)
 
-		return diag.FromErr(err)
+		return append(diagsReturn, diag.FromErr(err)...)
 	}
 	if err := sess.commitConf("update resource junos_security_ipsec_vpn", jnprSess); err != nil {
 		sess.configClear(jnprSess)
 
-		return diag.FromErr(err)
+		return append(diagsReturn, diag.FromErr(err)...)
 	}
 	d.Partial(false)
 
-	return resourceIpsecVpnRead(ctx, d, m)
+	return append(diagsReturn, resourceIpsecVpnRead(ctx, d, m)...)
 }
 func resourceIpsecVpnDelete(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	sess := m.(*Session)
@@ -327,7 +346,7 @@ func setIpsecVpn(d *schema.ResourceData, m interface{}, jnprSess *NetconfObject)
 	sess := m.(*Session)
 	configSet := make([]string, 0)
 
-	if d.Get("bind_interface_auto").(bool) {
+	if d.Get("bind_interface").(string) != "" {
 		configSet = append(configSet, "set interfaces "+d.Get("bind_interface").(string))
 	}
 	setPrefix := "set security ipsec vpn " + d.Get("name").(string)
@@ -503,9 +522,8 @@ func delIpsecVpn(d *schema.ResourceData, m interface{}, jnprSess *NetconfObject)
 	sess := m.(*Session)
 	configSet := make([]string, 0, 1)
 	configSet = append(configSet, "delete security ipsec vpn "+d.Get("name").(string))
-	if d.Get("bind_interface_auto").(bool) {
-		empty := checkInterfaceNC(d.Get("bind_interface").(string), m, jnprSess)
-		if empty == nil {
+	if d.Get("bind_interface").(string) != "" {
+		if empty := checkInterfaceNC(d.Get("bind_interface").(string), m, jnprSess); empty == nil {
 			configSet = append(configSet, "delete interfaces "+d.Get("bind_interface").(string))
 		}
 	}