+++ CHANGELOG.md
@@ -3,6 +3,7 @@ ENHANCEMENTS:
 * add `traffic_selector` argument in resource `security_ipsec_vpn` (Fixes [#53](https://github.com/jeremmfr/terraform-provider-junos/issues/53))
 
 BUG FIXES:
+* remove useless ForceNew for `bind_interface_auto` argument in resource `security_ipsec_vpn`
 
 ## v1.7.0
 ENHANCEMENTS: